% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{add_script}
\alias{add_script}
\alias{add_script_cleaning}
\alias{add_script_analysis}
\title{Add a script}
\usage{
add_script(path, template = NULL, open = TRUE)

add_script_cleaning(name)

add_script_analysis(name)
}
\arguments{
\item{path}{the path of the new file to create}

\item{template}{optional, the name of a template that the new file should copy. Must
be a pattern that uniquely matches one template name}

\item{open}{TRUE or FALSE - should the new script be opened in R Studio?}
}
\description{
\code{add_script()} creates a file (like\code{\link[=file.create]{file.create()}}), but it also enables the new file to be created from a template.
\code{add_script_cleaning()}, \code{add_script_analysis()}, and \code{add_script_helper()} are wrappers that automatically place
the new scripts in the proper "cleaning", "analysis", or "helper script" directories and use default templates. These
functions only work from inside an analysis directory.
}
