% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{analysis_directory_helpers}
\alias{analysis_directory_helpers}
\alias{peek_raw}
\alias{peek_cleaned}
\alias{read_raw}
\alias{read_cleaned}
\alias{write_raw}
\alias{write_cleaned}
\title{Work with "Raw Data" and "Cleaned Data" Inside an Analysis Folder}
\usage{
peek_raw()

peek_cleaned()

read_raw(file, ...)

read_cleaned(file, ..., is_rds = TRUE)

write_raw(x, file, na = "", ...)

write_cleaned(x, file, na = "", ...)
}
\arguments{
\item{file}{for "read_" functions, a pattern that uniquely matches one filename in "01 Raw Data" or "03 Cleaned Data".
for "write_" functions, then exact name of the output file (including file extension is optional)}

\item{...}{arguments to pass to \code{\link[readr:read_csv]{readr::read_csv()}} or \code{\link[readr:write_csv]{readr::write_csv()}}}

\item{is_rds}{TRUE or FALSE - should the ".rds" version be read? The recommended (and default)
values is TRUE. If FALSE, the ".csv" version of the cleaned data will be read}

\item{x}{a tibble}

\item{na}{for "write_" functions, the string used for missing values. By default this is "" (empty cell),
unlike the default for \code{write_csv()}, which is the string "NA"}
}
\description{
From anywhere inside of an analysis folder, these functions facilitate working with the
"01 Raw Data" and "03 Cleaned Data" subdirectories. \code{peek_raw()} and \code{peek_cleaned()} show
the available filenames within these directories. \code{read_raw()} and \code{read_cleaned()} allow
you to read an available file from these directories, and allow you to specify file by a
uniquely identifying regex pattern rather than a full filename (see \code{\link[=read_match]{read_match()}}).
\code{write_raw()} writes a tibble to "01 Raw Data" (as a csv by default).
\code{write_cleaned()} writes a tibble to "03 Cleaned Data" in both csv and rds formats.
}
\details{
Be careful when using \code{write_raw()} - you should only use when pulling in a data source and
immediately saving the result. Data with any further processing should not be saved in "01 Raw Data"

NOTE: you do not have to be in the top-level of an analysis directory to use these functions.
You can be ANYWHERE within an analysis folder and the function will correctly find the
"01 Raw Data" or "03 Cleaned Data" subdirectory. If your working directory is not within an
analysis folder, all of these functions should throw an error indicating that you've reached
the root of your system.
}
