% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{as_vec}
\alias{as_vec}
\alias{as_chr}
\alias{as_dbl}
\alias{as_int}
\alias{as_lgl}
\alias{as_fct}
\title{Convert A Vector or Flat List}
\usage{
as_vec(x, .ptype = NULL)

as_chr(x)

as_dbl(x)

as_int(x)

as_lgl(x)

as_fct(x, levels = NULL)
}
\arguments{
\item{x}{A vector or flat list}

\item{.ptype}{optional, the target ptype for the conversion}

\item{levels}{optional character vector specifying factor levels. If not provided, levels will be all
values that appear in \code{x}}
}
\value{
A vector
}
\description{
Given a vector or a flat list, this function converts to a vector of a given ptype. This is especially
useful for combining lists of length-1 elements into a vector. If no ptype provided, the conversion will
be to the common ptype.
}
\details{
The following variants enforce conversion to specific types of vectors:
\itemize{
\item as_chr() to character(0)
\item as_dbl() to double(0)
\item as_int() to integer(0)
\item as_lgl() to logical(0)
\item as_fct() to factor(0)
}
}
