% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_helpers.R
\name{assert_scalar}
\alias{assert_scalar}
\alias{assert_scalarish}
\alias{assert_scalar_character}
\title{Assert Scalar Types}
\usage{
assert_scalar(x, x_arg = rlang::caller_arg(x))

assert_scalarish(x, x_arg = rlang::caller_arg(x))

assert_scalar_character(x, x_arg = rlang::caller_arg(x))
}
\arguments{
\item{x}{An object}

\item{x_arg}{string to label \code{x} in the error message}
}
\value{
returns TRUE or throws an error
}
\description{
Assert Scalar Types
}
