% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{clear}
\alias{clear}
\alias{clear_all}
\alias{clear_except}
\title{Clear Variables from the Global Environment}
\usage{
clear()

clear_all()

clear_except(...)
}
\arguments{
\item{...}{For \code{clear_except()}, variables to retain in the global environment.}
}
\value{
No return value - updates the global environment
}
\description{
\code{clear()} clears all variables from the global environment, except for functions and database
connections. \code{clear_all()} removes everything, including functions and database connections.
\code{clear_except()} removes everything (including functions and database connections), except for
the objects specific in \code{...}
}
