% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{collect_sis}
\alias{collect_sis}
\title{Collect an SIS Table and Make More Consistent with Data Warehouse}
\usage{
collect_sis(x, ...)
}
\arguments{
\item{x}{A remote SIS table}

\item{...}{arguments passed to \code{\link[dplyr:collect]{dplyr::collect()}}}
}
\value{
A tibble
}
\description{
A wrapper for \code{\link[dplyr:collect]{dplyr::collect()}} that makes a couple of tweaks to make SIS source
data more consistent with warehouse data. Namely, it renames \code{ID} to \code{StudentID} and
ensures that it is a character vector. Also if \code{DispSem} exists it is renamed to
\code{StandardAcademicPeriod} and formatted from SIS style (e.g., "FL2020") to Data
Warehouse style (e.g., "Fall 2020")
}
