% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{connect_db}
\alias{connect_db}
\alias{connect_dw}
\alias{connect_sis}
\alias{disconnect_dw}
\alias{disconnect_sis}
\title{Database Connection Helpers}
\usage{
connect_db(server, database, uid = Sys.getenv("ODBC_UID"), conn_name = "conn")

connect_dw(uid = Sys.getenv("ODBC_UID"))

connect_sis(uid = Sys.getenv("ODBC_UID"))

disconnect_dw()

disconnect_sis()
}
\arguments{
\item{server, database}{strings giving the name of the server and database}

\item{uid}{your WashU email address. By default, the function retrieves this from ".Renviron". See \code{details}.}

\item{conn_name}{a string, "conn" by default. This will be the variable name for the new connection in the global environment.}
}
\value{
Technically returns only \code{TRUE} if successful. It creates a database connection and assigns it to \code{conn_name} in the global
environment.
}
\description{
\code{connect_db()} connects to a database, given a \code{server} and \code{database}. After calling the function, you will need to authenticate
interactively by providing your WashU credentials. \code{connect_db()} and \code{connect_sis()} are convenient wrappers that connect to the
data warehouse and SIS archive, respectively. DO NOT assign the result - the connection will automatically be assigned a standard
name (\code{connect_dw()} creates a connection named \code{conn_dw} and \code{connect_sis()} creates a connection named \code{conn_sis}).
\code{disconnect_dw()} and \code{disconned_sis()} disconnect from the warehouse and sis archive, respectively, and remove the names of the
connections from the environment.
}
\details{
If the \code{uid} is not provided (or not valid), the function will still work. You will just need to go through an extra Microsoft Sign in
screen in which you provide your WashU email. The easiest way to always skip this step (and not manually enter your email as
the \code{uid} argument) is to store your email as the "ODBC_UID" parameter in the ".Renviron" file. To do that, open the ".Renviron"
file by calling \code{usethis::edit_r_environ()}. Then add this line to the file: ODBC_UID = "youremail@wustl.edu".
Save the ".Renvion" file and then restart R for the change to take effect.
}
