% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{database_helpers}
\alias{database_helpers}
\alias{db_names}
\alias{db_nrow}
\alias{db_col_vals}
\alias{db_count}
\title{Database Helpers}
\usage{
db_names(tbl)

db_nrow(tbl)

db_col_vals(tbl, col)

db_count(tbl, ..., wt = NULL, sort = FALSE, name = NULL)
}
\arguments{
\item{tbl}{A \code{tbl} object linked to the table of a database}

\item{col}{Column name as a string or symbol}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Variables to group
by.}

\item{wt}{<\code{\link[rlang:args_data_masking]{data-masking}}> Frequency weights.
Can be \code{NULL} or a variable:
\itemize{
\item If \code{NULL} (the default), counts the number of rows in each group.
\item If a variable, computes \code{sum(wt)} for each group.
}}

\item{sort}{If \code{TRUE}, will show the largest groups at the top.}

\item{name}{The name of the new column in the output.

If omitted, it will default to \code{n}. If there's already a column called \code{n},
it will use \code{nn}. If there's a column called \code{n} and \code{nn}, it'll use
\code{nnn}, and so on, adding \code{n}s until it gets a new name.}
}
\description{
Helpers to get information about the current state of a \code{tbl} without using \code{collect()} to
pull it into a tibble/data frame.
\itemize{
\item \code{db_names()} gives the column names (alias for \code{colnames()})
\item \code{db_nrow()} gives the number of rows
\item \code{db_col_vals()} shows all the unique values in a given column
\item \code{db_count()} works just like \code{\link[dplyr:count]{dplyr::count()}} for remote tables, but returns as a data frame
(rather than \code{tbl}) to facilitate printing of the full counts table
}
}
