% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots_helpers.R
\name{dots_flat}
\alias{dots_flat}
\alias{dots_chr}
\title{Process \code{...} Arguments as a Flat List or Vector}
\usage{
dots_flat(..., .ptype = NULL)

dots_chr(...)
}
\arguments{
\item{...}{A vector, a flat list, or a series of scalar arguments}

\item{.ptype}{Prototype giving the desired type for the output vector. If \code{NULL} (the default), the
output will be a flat list}
}
\value{
A vector or flat list
}
\description{
This function accepts two different forms of \code{...} and ensures that the input is processed as a flat list or vector.
\code{...} can be either a single vector/list or a series of scalar elements. If NOT a single vector/list, then all
elements must be scalar, and the individual elements will be compiled into a vector or flat list.
Variants produce specified output types:
\itemize{
\item \code{dots_chr()} returns a character vector
}
}
\details{
Formulas are considered scalar in the context of \code{\link[=dots_flat]{dots_flat()}}. See \code{\link[=is_scalar]{is_scalar()}}.
}
