% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{dw_tbl_census}
\alias{dw_tbl_census}
\alias{dw_tbl_AcademicRecord}
\alias{dw_tbl_AcademicPeriodRecord}
\alias{dw_tbl_ProgramOfStudyRecord}
\alias{dw_tbl_RegistrationRecord}
\alias{dw_tbl_SectionRoleAssignment}
\title{Retrieve a Census Package from the Data Warehouse}
\usage{
dw_tbl_census(pkg, sems = "Fall 2025", weeks = 10, conn = conn_dw)

dw_tbl_AcademicRecord(sems = "Fall 2025", weeks = 10, conn = conn_dw)

dw_tbl_AcademicPeriodRecord(sems = "Fall 2025", weeks = 10, conn = conn_dw)

dw_tbl_ProgramOfStudyRecord(sems = "Fall 2025", weeks = 10, conn = conn_dw)

dw_tbl_RegistrationRecord(sems = "Fall 2025", weeks = 10, conn = conn_dw)

dw_tbl_SectionRoleAssignment(sems = "Fall 2025", weeks = 10, conn = conn_dw)
}
\arguments{
\item{pkg}{pattern matching the name of exactly table in the "STUCENSUS" schema}

\item{sems}{character vector giving the semesters to include. This is fairly flexible, but each element
must include text that's interpretable as either "Fall" or "Spring" and a year (either 2- or 4-digits).}

\item{weeks}{character or numeric vector giving the weeks to include. Can include only 0, 4, or 10
(or string variants such as "0", "4", "10" or "00", "04", "10").}

\item{conn}{a database connection to the the SIS archive.}
}
\value{
A \code{tbl()} object.
}
\description{
Creates a \code{tbl} object linked to a given census package. You can interact with this object via dplyr verbs as
needed and then call \code{\link[dplyr:collect]{dplyr::collect()}} to pull the result into a data frame. Commonly, you will use
\code{filter()} and \code{select()} to narrow the data before using \code{collect()}.
By default, this will filter the data to the Fall 2025 10th week census, but this can be adjusted to include any census
or censuses of interest. These are convenient wrappers to link to specific packages:
\itemize{
\item \code{dw_tbl_AcademicRecord()} for "vAcademicRecord"
\item \code{dw_tbl_AcademicPeriodRecord()} for "vAcademicPeriodRecord"
\item \code{dw_tbl_ProgramOfStudyRecord()} for "vProgramOfStudyRecord"
\item \code{dw_tbl_RegistrationRecord()} for "vRegistrationRecord"
\item \code{dw_tbl_SectionRoleAssignment()} for "vSectionRoleAssignment"
}
}
