% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{dw_tbl_student}
\alias{dw_tbl_student}
\alias{dw_tbl_student_AcademicRecord}
\alias{dw_tbl_student_AcademicPeriodRecord}
\alias{dw_tbl_student_CourseSectionDefinition}
\alias{dw_tbl_student_ProgramOfStudyDefinition}
\alias{dw_tbl_student_RegistrationRecord}
\alias{dw_tbl_student_WaitlistUtilization}
\title{Retrieve a Student Data Package from the Warehouse}
\usage{
dw_tbl_student(pkg, conn = conn_dw)

dw_tbl_student_AcademicRecord(conn = conn_dw)

dw_tbl_student_AcademicPeriodRecord(conn = conn_dw)

dw_tbl_student_CourseSectionDefinition(conn = conn_dw)

dw_tbl_student_ProgramOfStudyDefinition(conn = conn_dw)

dw_tbl_student_RegistrationRecord(conn = conn_dw)

dw_tbl_student_WaitlistUtilization(conn = conn_dw)
}
\arguments{
\item{pkg}{pattern matching the name of exactly table in the "STUDENT" schema}

\item{conn}{a database connection to the the data warehouse (called \code{conn_dw} by default)}
}
\value{
A \code{tbl} object
}
\description{
Creates a \code{tbl} object linked to a given student data package (Workday daily refresh tables). You can
interact with this object via \code{dplyr} verbs as needed and then call \code{\link[dplyr:collect]{dplyr::collect()}} to pull the result
into a data frame. Commonly, you will use \code{filter()} and \code{select()} to narrow the data before using \code{ collect()}.
These are convenient wrappers to link to specific packages:
\itemize{
\item \code{dw_tbl_student_AcademicRecord()} for "cvAcademicRecord"
\item \code{dw_tbl_student_AcademicPeriodRecord()} for "cvAcademicPeriodRecord"
\item \code{dw_tbl_student_CourseSectionDefinition()} for "cvCourseSectionDefinitionStudyRecord"
\item \code{dw_tbl_student_ProgramOfStudyDefinition()} for "cvProgramOfStudyDefinition"
\item \code{dw_tbl_student_RegistrationRecord()} for "cvRegistrationRecord"
\item \code{dw_tbl_student_WaitlistUtilization()} for "cvWaitlistUtilization"
}
}
