% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{ensure_extension}
\alias{ensure_extension}
\alias{ensure_r}
\alias{ensure_rmd}
\alias{ensure_rds}
\alias{ensure_csv}
\title{Ensure Filename has a Given Extension}
\usage{
ensure_extension(file, ext)

ensure_r(file)

ensure_rmd(file)

ensure_rds(file)

ensure_csv(file)
}
\arguments{
\item{file}{A filename}

\item{ext}{A file extension (will work with our without the leading ".")}
}
\value{
a filename as a string
}
\description{
Checks if a filename has an expected extension and if not, adds it. Other \code{ensure_} functions
are wrappers that check for a specific extention (e.g., \code{ensure_r()} checks for a ".R" extension)
}
