% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_year}
\alias{extract_year}
\title{Extract Year}
\usage{
extract_year(x, is_2000 = TRUE)
}
\arguments{
\item{sem}{A vector containing semesters in some format}
}
\value{
An integer vector of 4-digit years
}
\description{
Given any input vector that consistently contains 2-digit and/or 4-digit numbers, this function extracts into
a vector of "years". 2-digit numbers are automatically prepenced with "20".
}
