% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{fetch_programInfo}
\alias{fetch_programInfo}
\alias{fetch_programInfo_sis}
\alias{fetch_programInfo_workday}
\title{Retrieve SIS and Workday Program of Study Inventories}
\usage{
fetch_programInfo_sis()

fetch_programInfo_workday()
}
\value{
A tibble
}
\description{
These functions retrieve program inventory files from "OO IR Office/Data Sources, Resources/SOURCE-Workday Student/Program Inventory".
These files were pulled from the source data (SIS Archive "htv_programs" table and Workday report SRPT0027) on 8/4/2026 and and slightly updated
(remove, rearrange columns and added columns to facilitate linking programs across SIS and Workday). One column in the SIS data (ProgramType.revised)
was cleaned to better match Workday categories, but the original column is also retained.
}
