% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{fetch_sis}
\alias{fetch_sis}
\alias{fetch_progHist_sis}
\alias{fetch_progHist_dw}
\alias{fetch_degrees_sis}
\title{Fetch Useful Datasets from Data Warehouse or SIS}
\usage{
fetch_progHist_sis(sems = f13:sp25, add_program_info = TRUE)

fetch_progHist_dw(sems = "Fall 2025", weeks = 10, conn = conn_dw)

fetch_degrees_sis(sems = f13:su25, add_program_info = TRUE)
}
\arguments{
\item{sems}{semesters to include. Can include consecutive semesters with ":" syntax}

\item{add_program_info}{TRUE or FALSE - should supplementary program info columns be added?}
}
\value{
A tibble
}
\description{
Whereas \code{sis_dw_} and \code{sis_tbl_} functions connect to tables as-is and return a \code{tbl} connection object, \code{fetch_} functions
are designed to return more "analysis ready" tibbles. They do things such as convert SIS data to data warehouse format
(e.g., making column names and date formats consistent), join in supplementary data, order columns based on importance /
frequencey of use, and perform other data manipulation to output data that is ready-made for a specific purpose.
}
\details{
\itemize{
\item \code{fetch_progHist_sis()} pulls in a transformed version of "htv_student_prog_hist"
\item \code{fetch_degrees_sis()} pulls in a transformed version of "htv_degree_info"
}
}
