% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{file_extension}
\alias{file_extension}
\alias{file_basename}
\title{Easily Extract Pieces of a Filename}
\usage{
file_extension(file)

file_basename(file)
}
\arguments{
\item{x}{A filename or vector of filenames}
}
\value{
A string - the requested piece of \code{filename}
}
\description{
\code{file_extension()} extracts the file extension (including the leading ".") from a filename, and
\code{file_basename()} removes the extension
}
