% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{file_path}
\alias{file_path}
\title{Return normalized path}
\usage{
file_path(...)
}
\arguments{
\item{...}{path components}
}
\value{
full absolute path
}
\description{
Wrapper for \code{\link[=file.path]{file.path()}} that returns a normalized path
}
