% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{filter_sems}
\alias{filter_sems}
\alias{filter_sems_sis}
\title{Filter a Data Warehouse or SIS Table by Semester}
\usage{
filter_sems(x, ..., to_sis = FALSE, include_summer = FALSE, sem_col = waiver())

filter_sems_sis(x, ..., include_summer = FALSE, sem_col = DispSem)
}
\arguments{
\item{x}{Data Warehouse or SIS Data as a remote table or dataframe}

\item{...}{semesters entered as strings, symbols, or character vectors. \code{:} expressions will be expanded
to include all consecutive semesters (e.g., \code{fl20:sp22} would include "Fall 2020", "Spring 2021", "Fall 2021", and "Spring 2022")}

\item{to_sis}{TRUE or FALSE - should the output be in SIS format?}

\item{include_summer}{should summer semesters be included when expanding consecutive semesters?
\code{FALSE} by default since summer semesters are not of interest in most analyses}

\item{sem_col}{the name of the semester column as a string or symbol. Defaults guesses are made based
on the \code{to_sis} argument}
}
\value{
A tibble or remote table (same as \code{x})
}
\description{
Filter an SIS or data warehouse table by semester. \code{...} arguments are processed with \code{\link[=sems_expand]{sems_expand()}},
allowing flexible and efficient generation of many semsters. \code{filter_sems_sis()} is just a wrapper for
\code{filter_sems()} with \code{to_sis} set to TRUE.
}
