% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{go_cur_file}
\alias{go_cur_file}
\title{Set Working Directory to Wherever the Current Script is}
\usage{
go_cur_file()
}
\value{
no return, changes working directory
}
\description{
Set Working Directory to Wherever the Current Script is
}
