% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{helper_scripts}
\alias{helper_scripts}
\alias{peek_helpers}
\alias{add_script_helper}
\alias{source_helper}
\alias{edit_helper}
\title{Source a Helper File}
\usage{
peek_helpers()

add_script_helper(name)

source_helper(name)

edit_helper(name)
}
\arguments{
\item{name}{For \code{add_script_helper()}, the name of the ".R" script to create in "Helper Scripts". For
\code{source_helper()} and \code{edit_helper()}, a pattern that uniquely matches exactly one script in
"Helper Scripts"}
}
\description{
Helper functions that facilitates interacting with the "Helper Scripts" subdirectory within an analysis
folder. You must be somewhere inside of an analysis folder, for these functions to work.
\itemize{
\item \code{peek_helpers()} shows all available helper files in "Helper Scripts"
\item \code{add_script_helper()} adds a script to "Helper Scripts"
\item \code{source_helper()} sources a script from "Helper Scripts"
\item \code{edit_helper()} opens a script from "Helper Scripts" for editing
}
}
