% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{hook}
\alias{hook}
\title{Use a Pattern to "hook" the Matching Value}
\usage{
hook(pattern, table, empty = c("error", "return_pattern", "return_na"))
}
\arguments{
\item{pattern}{a single regex pattern}

\item{table}{a vector of candidate values}

\item{empty}{Controls what happens when no matches occur. By default ("error") an error is thrown.
You can also choose to return \code{pattern} or return \code{NA}}
}
\value{
A single string
}
\description{
Given a pattern and set of candidate values, this function returns the single matching value. This is
designed to be helper function that facilitates users' ability to use a (short) regex pattern as a
stand-in for a (possibly long) full value (e.g. a filename). It provides safety by ensuring that the
pattern "hooks" a single value (presumably the one the user intended the pattern to represent.)
This function always throws an error when \code{pattern} matches more than one value in \code{vals}. Use the
\code{empty} argument to control what happens when there are no matches.
}
\details{
This function is really just a wrapper for \code{\link[=str_subset1]{str_subset1()}} with the arguments renamed and
rearranged to more intuitively match the use case (using a single pattern to retrieve a
single candidate value)
}
