% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{hook_file}
\alias{hook_file}
\alias{hook_path}
\title{Return the Filename that Matches a Pattern}
\usage{
hook_file(pattern, dir_path = ".", return_full_path = FALSE)

hook_path(pattern, dir_path)
}
\arguments{
\item{pattern}{A pattern that matches exactly one filename in dir_path}

\item{dir_path}{path giving the directory in which to search for \code{pattern}. Defaults to working directory}

\item{return_full_path}{TRUE or FALSE - should the full path be returned? If FALSE, just the filename is returned}
}
\value{
A filename, or if \code{return_full_path} is TRUE, then a full path
}
\description{
\code{hook_file()} is a helper that "hooks" a single filename based on a regex pattern (see \code{\link[=hook]{hook()}}). Use
inside of read functions to allow those functions to accept patterns rather than full filenames. \code{hook_path()}
is just a wrapper that returns the full path.
}
