% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iehelpr-package.R
\docType{package}
\name{iehelpr-package}
\alias{iehelpr}
\alias{iehelpr-package}
\title{iehelpr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cahillmj/iehelpr}
  \item Report bugs at \url{https://github.com/cahillmj/iehelpr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
