% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_characterish}
\alias{is_characterish}
\title{Is an object characterish?}
\usage{
is_characterish(x)
}
\arguments{
\item{x}{A vector}
}
\value{
TRUE or FALSE
}
\description{
Is an object either character or factor?
}
