% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_evaluable}
\alias{is_evaluable}
\title{Is an Expression or Quosure evaluable?}
\usage{
is_evaluable(expr)
}
\arguments{
\item{expr}{An expression or quosure}
}
\value{
TRUE or FALSE
}
\description{
Tests whether an expression or quosure can be evaluated with \code{\link[rlang:eval_tidy]{rlang::eval_tidy()}}.
}
