% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_length}
\alias{is_length}
\title{Check Object Length}
\usage{
is_length(x, len)
}
\arguments{
\item{x}{An r object}

\item{len}{the target length to check for}
}
\value{
TRUE or FALSE
}
\description{
Convenience function to check length (e.g., is_length(x, 2) vs identical(length(x), 2L)).
}
