% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_scalarish}
\alias{is_scalarish}
\title{Is an Object Scalar or a Formula}
\usage{
is_scalarish(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE or FALSE
}
\description{
This test expands the definition of "scalar" to include a single formula. Even though formulas are
technically length 2 or 3, it often makes sense to think of them as single element. \code{is_scalarish()}
facilitates this by testing whether something is either a "real scalar" or a formula.
}
