% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_setdiff}
\alias{is_setdiff}
\title{Is There a Set Difference?}
\usage{
is_setdiff(x, y)
}
\arguments{
\item{x, }{y vectors}
}
\value{
TRUE or FALSE
}
\description{
Test whether there is a setdiff. Returns \code{TRUE} if \code{x} has any items that are not in \code{y}
}
