% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{length_checks}
\alias{length_checks}
\alias{not_empty}
\alias{is_scalar}
\alias{is_multiple}
\title{Object Length Checks}
\usage{
not_empty(x)

is_scalar(x)

is_multiple(x)
}
\arguments{
\item{x}{An R object}
}
\value{
TRUE or FALSE
}
\description{
Convenience functions to check important length thresholds, complementing \code{\link[rlang:is_empty]{rlang::is_empty()}}
\itemize{
\item \code{\link[rlang:is_empty]{rlang::is_empty()}} - is the length 0?
\item \code{not_empty()} - is the length > 0
\item \code{is_scalar()} - is the length exactly 1?
\item \code{is_multiple()} - is the length > 1?
}
}
