% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_helpers.R
\name{meta}
\alias{meta}
\title{Get a Useful Summary of a Data Frame's Fields}
\usage{
meta(x, exclude_num_levels = TRUE, max_levels = 10, n_values = 4)
}
\arguments{
\item{x}{A data frame}

\item{exclude_num_levels}{logical flag, should the \code{levels} column "exclude" numeric fields? If \code{TRUE}, then
the \code{levels} value will be \code{NA} for all of the numeric columns in \code{x}.}

\item{max_levels}{an integer, the maximum number of category levels to display. Levels will only be displayed
for columns with a number of distinct values less than or equal to this number. The \code{levels} value will be \code{NA}
for columns in \code{x} that have more distinct values than this.}

\item{n_values}{an integer, the number of values to display in the \code{values} column.}
}
\value{
A data frame - a summary of \code{x} in which every row describes a column of \code{x}
}
\description{
Similar to \code{\link[dplyr:glimpse]{dplyr::glimpse()}}, this function returns information (name, type, levels, sample values)
about every field in the data frame.
}
