% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{peek_database}
\alias{peek_database}
\alias{dw_peek_census}
\alias{sis_peek_htv}
\title{See Available Tables}
\usage{
dw_peek_census()

sis_peek_htv()
}
\value{
A character vector of table names
}
\description{
\code{dw_peek_census()} shows the names of all available packages in the "STUCENSUS" schema of the data warehouse.
\code{sis_peek_htv()} shows the names of all available SIS tables with an "htv_" prefix. NOTE: the appropriate
database connection (\code{conn_dw} or \code{conn_sis}) must be active
}
