% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{programs_harmonize}
\alias{programs_harmonize}
\title{Format SIS Program Data in Data Warehouse Format}
\usage{
programs_harmonize(x)
}
\arguments{
\item{x}{A character vector of SIS program names}
}
\value{
A character vector
}
\description{
Convert programs from SIS format (e.g., B.S. MAJOR IN CHEMISTRY) to data warehouse format
(e.g., Chemistry, B.S.). Note this does not directly link old SIS programs data warehouse programs,
it merely make the programs follow data warehouse format conventions (e.g., title case, degree at
the end preceded by a comma). If possible you should link SIS programs directly to their
data warehouse counterparts, e.g., by calling \code{fetch_programInfo_sis()} and using the \code{ProgName}
and \code{ProgName.warehouse} fields. The main purpose of \code{programs_harmonize} is to update the
formatting of SIS programs that you aren't able to match directly.
}
