% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{quo_is_waiver}
\alias{quo_is_waiver}
\title{Check Whether a Quosure is Storing a Waiver Object}
\usage{
quo_is_waiver(quo)
}
\arguments{
\item{quo}{A quosure or expression}
}
\value{
TRUE or FALSE
}
\description{
Check whether an expression or quosure evaluates to \code{waiver()}. This is useful for arguments
captured with \code{\link[rlang:enexpr]{rlang::enexpr()}} or \code{\link[rlang:enquo]{rlang::enquo()}} to check whether the captured argument
was actually \code{\link[ggplot2:waiver]{waiver()}}
}
