% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{read_filepath}
\alias{read_filepath}
\alias{read_match}
\title{Convenience Functions to Facilitate Reading Data}
\usage{
read_filepath(file, ...)

read_match(filename, dir_path = ".", ...)
}
\arguments{
\item{file}{path to a file}

\item{...}{arguments to pass along to the underlying read function (the function used will
depend upon the file extension).}

\item{filename}{a pattern that matches exactly one filename in \code{dir_path}}

\item{dir_path}{path to the directory containing the data file. "." by default}
}
\value{
The result of reading the file, most likely a tibble
}
\description{
\code{read_filepath()} detects file extensions and reads the most common types of rectangular data
files (csv, excel, and rds). \code{read_match()} does the same and also allows you to specify the file with
a pattern that matches exactly one filename in \code{dir_path}. This allows you to avoid writing or copying
long filenames when reading data.
}
