% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_helpers.R
\name{recode_hook}
\alias{recode_hook}
\alias{replace_hook}
\title{Recode and Replace Using Regex Patterns}
\usage{
recode_hook(
  x,
  ...,
  from = NULL,
  to = NULL,
  default = NULL,
  unmatched = "default",
  ptype = NULL,
  enforce_match = FALSE
)

replace_hook(x, ..., from = NULL, to = NULL)
}
\arguments{
\item{x}{A vector.}

\item{...}{two sided formulas where \code{lhs} determines the values to be recoded and \code{rhs} provides the replacement
value. The \code{lhs} can be a character vector of any size, and each element must be a regex pattern that uniquely
matches a single value in \code{x}. \code{...} is mutually exclusive with \code{from} and \code{to}}

\item{from}{Values to look up in \code{x} and map to values in \code{to}.

Typically this is a single vector of any size that is
\link[vctrs:theory-faq-coercion]{cast} to the type of \code{x}. For more advanced
usage, this can be a list of vectors of any size each of which are
\link[vctrs:theory-faq-coercion]{cast} to the type of \code{x}.

Mutually exclusive with \code{...}.}

\item{to}{Values that \code{from} map to.

Typically this is a single vector that is
\link[vctrs:theory-faq-recycling]{recycled} to the size of \code{from}. For more
advanced usage, this can be a list of vectors each of which are
\link[vctrs:theory-faq-recycling]{recycled} to the size of \code{x}.

Mutually exclusive with \code{...}.}

\item{default}{Default value to use when there is a value present in \code{x}
that is unmatched by a value in \code{from}.

By default, a missing value is used as the default value.

If supplied, will be \link[vctrs:theory-faq-recycling]{recycled} to the size of
\code{x}.

Can only be set when \code{unmatched = "default"}.}

\item{unmatched}{Handling of unmatched locations.

One of:
\itemize{
\item \code{"default"} to use \code{default} in unmatched locations.
\item \code{"error"} to error when there are unmatched locations.
}}

\item{ptype}{An optional override for the output type, which is usually
computed as the common type of \code{to} and \code{default}.}

\item{enforce_match}{For \verb{recode_hook, }TRUE\code{or}FALSE\verb{. Should the function throw an error if there is a pattern in }...\code{or}from\verb{that doesn't match any values in}x\verb{. By default this is }FALSE\verb{- the function will inform the user of the unmatched pattern but still return the recoded version of}x`.}
}
\value{
A vector - \code{x} with specified updates
}
\description{
Wrappers for \code{\link[dplyr:recode_values]{dplyr::recode_values()}} and \code{\link[dplyr:replace_values]{dplyr::replace_values()}} that allow the user to use unique regex patterns
(rather than exact values) to specify which values to be replaced or recoded. Each pattern must uniquely match
exactly one value in the input vector (see \code{\link[=hook]{hook()}}). As with the original dplyr versions, the difference is that
\code{recode_hook()} intends to match and update every value in \code{x} (unmatched values become \code{NA} by default). \code{replace_hook()},
on the other hand, intends to update some specified values, but otherwise retains the original values of \code{x} that
aren't reference in \code{from} or \code{...}.
}
