% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_helpers.R
\name{replace_matches}
\alias{replace_matches}
\title{Replace all Values that Match Regex Patterns}
\usage{
replace_matches(x, ..., from = NULL, to = NULL)
}
\arguments{
\item{x}{a character vector}

\item{...}{two sided formulas where \code{lhs} determines the values to be recoded and \code{rhs} provides the replacement
value. The \code{lhs} must be a single pattern, and it can match any number of values in \code{x}. For each pattern, all
matching values in \code{x} will be replaced with the \code{rhs} value \code{...} is mutually exclusive with \code{from} and \code{to}}

\item{from}{regex pattern specifying values of \code{x} to replace}

\item{to}{Replacement values}
}
\value{
A character vector
}
\description{
Replace all items that match a pattern. Unlike \code{replace_hook()}, a single pattern does not need to match
a single value of \code{x}. A pattern can match any number of values in \code{x}, and all matching values will be replaced
with the same corresponding value.
}
