% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{sem_extract}
\alias{sem_extract}
\alias{extract_year}
\alias{extract_term}
\title{Extract Year or Term}
\usage{
extract_year(sem, is_2000 = TRUE)

extract_term(sem, to_sis = FALSE)
}
\arguments{
\item{sem}{A vector containing semesters. This is failry flexbile but all inputs must have text
intepretatable as "Fall", "Spring", or "Summer" (e.g., "Fall", "F", "Spring", "sp") and must
include a valid 2-digit or 4-digit year.}

\item{is_2000}{Should all dates be assumed to start with "20"? If \code{FALSE}, then all inputs
will need to include 4-digit years}
}
\value{
For \code{extract_year()}, an integer vector of 4-digit years. For \code{extract_term()} a
character vector with only the values "Fall", "Spring" and "Summer"
}
\description{
Given a vector of semesters (e.g., "Fall 2020", "Spring 2023", "FL25), \code{extract_year()} and
\code{extract_term()} extract the calendar year and term (e.g., "Fall" or "Spring"), respectively.
Semester inputs can be formatted flexibly as long as they are unambiguous.
}
