% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{semester_computations}
\alias{semester_computations}
\alias{compute_AY_startYear}
\alias{compute_AY}
\alias{compute_student_year}
\alias{compute_student_sem}
\title{Compute Useful Variables from Semesters}
\usage{
compute_AY_startYear(sem)

compute_AY(sem, prefix = "AY ")

compute_student_year(sem, entry_sem)

compute_student_sem(sem, entry_sem)
}
\arguments{
\item{sem}{A vector of semesters (e.g., StandardAcademicPeriod)}

\item{prefix}{A string to add as a prefix to the output. By default, this is "AY " (the space is intentional),
and the output will be formatted like "AY 2020-21". Set prefix to "" or \code{NULL} to return just "2020-21", set
it to "AY" (no space) to return "AY2020-21", or use the \code{prefix} argument to set any other custom prefix}

\item{entry_sem}{A vector representing entry semesters (e.g, cohort)}
}
\value{
A numeric vector
}
\description{
Given a vector of semesters, these functions compute useful variables.
}
\details{
\code{compute_AY} returns the academic year in typical text format (e.g., "AY 2024-25"). \code{compute_AY_startYear()}
computes the starting year for the academic year (e.g., 2020 for both "Fall 2020" and "Spring 2021").
Given both a vector a semesters and a second vector representing students' entry semesters,
\code{compute_student_year()} and \code{compute_student_semester()} calculate the students' year number (e.g.,
a student in their first year gets a \code{1}) and semester number (e.g., a student in their third semesters
gets a \code{3}), respectively.

The primary reason to use \code{compute_AY_startYear()} is when you want to use the output in further calculations.
Otherwise, \code{compute_AY()} is usually preferable, especially if the output could be used in labels
(e.g., in a plot or table).

For \code{compute_student_semester()}, summer semesters are counted as semester number 2.5 of the academic year.
That way, every year counts as 2 semesters (e.g., A student's second Fall is their third semester), but there
is still a way to represent summer semesters.
}
