% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{sems_expand}
\alias{sems_expand}
\alias{sems_expand_sis}
\title{Expand Semesters that Include ":" Inputs}
\usage{
sems_expand(..., to_sis = FALSE, include_summer = FALSE)

sems_expand_sis(..., include_summer = FALSE)
}
\arguments{
\item{...}{semesters entered as strings, symbols, or character vectors. \code{:} expressions will be expanded
to include all consecutive semesters (e.g., \code{fl20:sp22} would include "Fall 2020", "Spring 2021", "Fall 2021", and "Spring 2022")}

\item{to_sis}{TRUE or FALSE - should the output be in SIS format?}

\item{include_summer}{should summer semesters be included when expanding consecutive semesters?
\code{FALSE} by default since summer semesters are not of interest in most analyses}
}
\value{
A character vector of semesters in the desired format
}
\description{
This function allows the use of \code{:} to refer to consecutive semesters. (e.g. FL23:SP25 to refer
to "Fall 2023", "Spring 2024", "Fall 2025", and "Spring 2025"). \code{sems_expand_sis()} is just a
convenient wrapper where \code{to_sis} is \code{TRUE}
}
