% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{sems_format}
\alias{sems_format}
\alias{sems_format_sis}
\title{Convert Semester Values to Proper Format}
\usage{
sems_format(..., to_sis = FALSE)

sems_format_sis(...)
}
\arguments{
\item{...}{Either a character vector or individual string arguments}

\item{to_sis}{TRUE or FALSE - should the output be formatted to SIS format?}
}
\value{
A character vector
}
\description{
Given fairly flexible inputs (e.g., "fl23"), \code{sems_format()} converts to proper semester
format for the data warehouse ("Fall 2023") or, if\code{to_sis} is TRUE, SIS ("FL2023"). \code{sems_format_sis()}
is just a convenient wrapper where \code{to_sis} is TRUE.
}
