% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{sems_harmonize}
\alias{sems_harmonize}
\title{Convert Semester Data from SIS Format to Data Warehouse Format}
\usage{
sems_harmonize(x)
}
\arguments{
\item{x}{a vector of semester data from an SIS table}
}
\value{
A character vector with semesters in data warehouse format
}
\description{
Convert semesters from SIS format (e.g., "FL2020", "SP2020") to data warehouse
format (e.g., "Fall 2020", "Spring 2020")
}
