% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{sis_tbl_htv}
\alias{sis_tbl_htv}
\alias{sis_tbl_degree}
\alias{sis_tbl_aa}
\alias{sis_tbl_milestones}
\alias{sis_tbl_progHist}
\alias{sis_tbl_divHist}
\alias{sis_tbl_demos}
\alias{sis_tbl_scores}
\alias{sis_tbl_courseRecord}
\title{Retrieve an "htv_" table from the SIS Archive}
\usage{
sis_tbl_htv(table, conn = conn_sis)

sis_tbl_degree(conn = conn_sis)

sis_tbl_aa(conn = conn_sis)

sis_tbl_milestones(conn = conn_sis)

sis_tbl_progHist(conn = conn_sis)

sis_tbl_divHist(conn = conn_sis)

sis_tbl_demos(conn = conn_sis)

sis_tbl_scores(conn = conn_sis)

sis_tbl_courseRecord(conn = conn_sis)
}
\arguments{
\item{table}{pattern matching the name of exactly one "htv_" table in the SIS archive}

\item{conn}{a database connection to the the SIS archive.}
}
\value{
A \code{tbl()} object.
}
\description{
Creates a \code{tbl} object linked to a given "htv_" table. You can interact with this object via dplyr verbs as
needed and then call \code{\link[dplyr:collect]{dplyr::collect()}} to pull the result into a data frame. Commonly, you will use
\code{filter()} and \code{select()} to narrow the data before using \code{collect()}.
These are convenient wrappers to link to commonly used tables:
\itemize{
\item \code{sis_tbl_degree()} for "htv_degree_info"
\item \code{sis_tbl_aa()} for "htv_academic_action"
\item \code{sis_tbl_milestones()} for "htv_milestones"
\item \code{sis_tbl_progHist()} for "student_prog_hist"
\item \code{sis_tbl_divHist()} for "stdt_div_hist"
\item \code{sis_tbl_demos()} for "stdt_demographics"
\item \code{sis_tbl_scores()} for "htv_scores"
\item \code{sis_tbl_courseRecord()} for "htv_stdt_course_rec"
}
}
