% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_helpers.R
\name{str_replace_multi}
\alias{str_replace_multi}
\title{Perform Multiple String Replacements at Once}
\usage{
str_replace_multi(string, ...)
}
\arguments{
\item{string}{A character vector}

\item{...}{Formulas where the \code{lhs} is a regex pattern, and \code{rhs} is the replacement value.}
}
\value{
A character vector
}
\description{
\code{str_replace_multi()} provides a compact way to perform multiple string replacements on the same
vector (as opposed to calling \code{\link[stringr:str_replace_all]{stringr::str_replace_all()}} multiple times to make these replacements).
Additionally, this function always detects matches on the original input vector, preventing accidental
matches that could occur with multiple calls to \code{str_replace_all}, where later calls are detecting
matches on updated versions of the vector.
}
