% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{str_subset1}
\alias{str_subset1}
\title{Attempt to Return a Single regex match}
\usage{
str_subset1(string, pattern, empty = c("error", "return_pattern", "return_na"))
}
\arguments{
\item{string}{A character vector}

\item{pattern}{A single regex pattern}

\item{empty}{Controls what happens when no matches occur. By default ("error") an error is thrown.
You can also choose to return \code{pattern} or return \code{NA}}
}
\value{
A single string
}
\description{
A variant of \code{\link[stringr:str_subset]{stringr::str_subset()}} to use when you want or expect to return a single string that
matches \code{pattern}.
This function always throws an error when more than one value of \code{string} matches \code{pattern}. Use the
\code{empty} argument to control what happens when there are no matches.
}
