% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helpers.R
\name{template_helpers}
\alias{template_helpers}
\alias{peek_templates}
\alias{copy_template}
\alias{edit_template}
\title{Work with Templates}
\usage{
peek_templates()

copy_template(template_name, to)

edit_template(template_name)
}
\arguments{
\item{template_name}{a pattern that uniquely matches the name of one template}

\item{to}{the path to which the template should be copied}
}
\description{
See and worth with templates in "00 IE/Resources/R Resources/Templates". \code{peek_templates()}
lets you see names of available templates, \code{copy_template()} creates a new file that copies
the template, and \code{edit_template()} opens the template in its source location for editing.
}
