% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{equal}
\alias{equal}
\title{More Lenient version of \code{identical()}}
\usage{
equal(x, y)
}
\arguments{
\item{x, y}{R objects}
}
\value{
TRUE or FALSE
}
\description{
\code{identical()} is very strict, requiring all classes and attributes to the exactly the same. \code{equal()} is a
slightly more lenient version that first attempts the two objects to a common \link[vctrs:vctrs]{vctrs} ptype before testing
whether they are identical. \verb{\%is\%} is just a convenient inline version of \code{equal()}
}
