% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{fetch_grs}
\alias{fetch_grs}
\title{Read the "RDS file" provided annually by Ryan Croft}
\usage{
fetch_grs(year = waiver(), is_rds = TRUE)
}
\arguments{
\item{year}{which year to read. By default will read the most recent year available}

\item{is_rds}{TRUE or FALSE - should you read the RDS version of the file? This is
recommended if possible because it avoids any conversion errors}
}
\value{
A data frame
}
\description{
Read the "RDS file" provided annually by Ryan Croft
}
