% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{fetch_programInfo}
\alias{fetch_programInfo}
\alias{fetch_programInfo_sis}
\alias{fetch_programInfo_workday}
\title{Retrieve SIS and Workday Program of Study Inventories}
\usage{
fetch_programInfo_sis()

fetch_programInfo_workday()
}
\value{
A tibble
}
\description{
These functions retrieve full program inventories for SIS and the Data Warehouse. Because the SIS archive is
frozen, the program inventory has been saved to file, and \code{fetch_programInfo_sis()} retrieves and reads
this static file from "OO IR Office/Data Sources, Resources/SOURCE-SIS Archive". \code{fetch_programInfo_workday()}
reads live data from the data warehouse ("STUDENT"."cvProgramOfStudyDefinition"), since the workday program
info can be updated at any time.
}
