% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access_helpers.R
\name{sis_fetch_progHist}
\alias{sis_fetch_progHist}
\title{Fetch "htv_progHist" and Supporting Info}
\usage{
sis_fetch_progHist(sems = fl13:sp24, add_program_info = TRUE)
}
\arguments{
\item{sems}{semesters to include. Can include consecutive semesters with ":" syntax}

\item{add_program_info}{TRUE or FALSE - should supplementary program info columns be added?}
}
\value{
A tibble
}
\description{
This function automatically collects the "htv_student_prog_hist" table from the SIS archive, harmonizes it with the
data warehouse (via \code{\link[=collect_sis]{collect_sis()}}), and optionally adds in additional program info (see \code{\link[=fetch_programInfo_workday]{fetch_programInfo_workday()}}).
}
